Add files via upload
Plants pests and diseases can cause significant economic losses for farmers, and accurate identification and classification of these issues is crucial. Deep learning techniques, such as Convolutional Neural Networks (CNNs) and You Only Look Once (YOLO), have demonstrated promising results in correctly identifying and classifying plant pests and diseases. However, the accuracy of these models heavily relies on the quality and size of the training dataset. Gathering a large and diverse dataset of plant pests and diseases can be challenging due to the variation in the appearance of diseased plants and the high cost of data collection. To tackle this issue, a web portal has been developed to gather datasets of plant pests and diseases, making it simpler for researchers to contribute data and enhance the accuracy of these models. In addition, Generative Adversarial Networks (GANs) have been used to expand the current datasets by creating synthetic images that closely resemble real-world data. This has resulted in improved accuracy and robustness of CNN and YOLO models in identifying and classifying plant pests and diseases. Moreover, recent developments in using GANs for dataset enlargement in plant pests and disease classification have been reviewed, and the advantages and challenges were highlighted. Using GANs for dataset enlargement has the potential to generate diverse images with varying levels of severity, leading to improved accuracy and generalization of CNN and YOLO models. Overall, the web portal and the use of GANs for dataset enlargement is a promising area of research that has the potential to significantly improve the accuracy and generalization of machine learning models for plant pests and disease classification.